requirements.txt
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pytesseract>=0.3.10
Pillow>=9.5.0
pydub>=0.23.1
openai-whisper>=20231117
datasets>=2.14.0
accelerate>=0.20.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
nltk>=3.8.0
tqdm>=4.65.0
 requests>=2.31.0
python-dotenv>=1.0.0
shap>=0.42.0